move what is left in upgrade_legacy.go to repair.go
